backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Deep Learning
tensorflow==2.15.0
keras==2.15.0
Pillow==10.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Production
gunicorn==21.2.0
